Added a Hit Particle Effect and cleanup
I tried to clean up the project window a little and got rid of some empty folders that were duplicates of existing folders. I renamed one of the scene folders as game Scenes that have all the scenes to better distinguish it from others that were imported form asset packs. Explains the large amount of changes.